Move remaining options to layers
